20231128 구현 완료

+++ CAC/src/main/java/com/example/CAC/service/CacService.java
@@ -1,7 +1,6 @@
 package com.example.CAC.service;
 
 import com.example.CAC.entity.Cac;
-import com.example.CAC.entity.User;
 import com.example.CAC.entity.CacRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -10,8 +9,12 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
+import javax.servlet.http.HttpSession;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
+
+
 
 @Service
 @RequiredArgsConstructor
@@ -30,11 +33,33 @@ public class CacService {
         return this.cacRepository.findByTitleContaining(keyword, pageable);
     }
 
+    public List<Cac> getAvg(String keyword) {
+        return cacRepository.findByTitleContaining(keyword);
+    }
+
     public Cac write(Cac cac){
         return cacRepository.save(cac);
     }
 
-    public void boardDelete(Long id){
-        cacRepository.deleteById(id);
+    public String boardDelete(Long id, HttpSession session) {
+        String userId = String.valueOf(session.getAttribute("user_id"));
+        Optional<Cac> cacOptional = cacRepository.findById(id);
+
+        if (cacOptional.isPresent()) {
+            Cac cac = cacOptional.get();
+            String kakaoId = String.valueOf(cac.getKakao_id());
+
+            if (kakaoId.equals(userId)) {
+                cacRepository.deleteById(id);
+                System.out.println("삭제 되었습니다.");
+                return "success";
+            } else {
+                System.out.println("권한이 없습니다.");
+                return "fail";
+            }
+        } else {
+            System.out.println("no entity");
+            return "error";
+        }
     }
 }
